Move more utilities over
