+++ packages/kbn-discover-utils/src/hooks/index.ts
@@ -6,10 +6,4 @@
  * Side Public License, v 1.
  */
 
-export * from './build_data_record';
-export * from './format_hit';
-export * from './format_value';
-export * from './get_doc_id';
-export * from './get_ignored_reason';
-export * from './get_should_show_field_handler';
-export * from './nested_fields';
+export * from './use_pager';
