+++ packages/kbn-discover-utils/src/utils/nested_fields.ts
