+++ src/plugins/discover/public/application/main/components/sidebar/lib/get_field_list.ts
@@ -10,7 +10,7 @@ import { difference } from 'lodash';
 import { type DataView, DataViewField } from '@kbn/data-views-plugin/public';
 import type { DatatableColumn } from '@kbn/expressions-plugin/common';
 import { fieldWildcardFilter } from '@kbn/kibana-utils-plugin/public';
-import { isNestedFieldParent } from '../../../utils/nested_fields';
+import { isNestedFieldParent } from '@kbn/discover-utils';
 
 export function getDataViewFieldList(
   dataView: DataView | undefined | null,
