+++ src/plugins/discover/public/components/doc_table/doc_table_embeddable.tsx
@@ -10,8 +10,8 @@ import React, { memo, useCallback, useMemo, useRef } from 'react';
 import './index.scss';
 import { FormattedMessage } from '@kbn/i18n-react';
 import { EuiText } from '@elastic/eui';
+import { usePager } from '@kbn/discover-utils';
 import { SAMPLE_SIZE_SETTING } from '../../../common';
-import { usePager } from '../../hooks/use_pager';
 import {
   ToolBarPagination,
   MAX_ROWS_PER_PAGE_OPTION,
