+++ src/plugins/discover/public/services/doc_views/components/doc_viewer_table/table.tsx
@@ -30,13 +30,17 @@ import { FormattedMessage } from '@kbn/i18n-react';
 import { debounce } from 'lodash';
 import { Storage } from '@kbn/kibana-utils-plugin/public';
 import { getFieldIconType } from '@kbn/unified-field-list/src/utils/field_types/get_field_icon_type';
-import { formatFieldValue, getIgnoredReason, getShouldShowFieldHandler } from '@kbn/discover-utils';
+import {
+  formatFieldValue,
+  getIgnoredReason,
+  getShouldShowFieldHandler,
+  isNestedFieldParent,
+  usePager,
+} from '@kbn/discover-utils';
 import { useDiscoverServices } from '../../../../hooks/use_discover_services';
-import { usePager } from '../../../../hooks/use_pager';
 import { FieldName } from '../../../../components/field_name/field_name';
 import { SHOW_MULTIFIELDS } from '../../../../../common';
 import { DocViewRenderProps, FieldRecordLegacy } from '../../doc_views_types';
-import { isNestedFieldParent } from '../../../../application/main/utils/nested_fields';
 import { TableFieldValue } from './table_cell_value';
 import { TableActions } from './table_cell_actions';
